fixed issue with search results
